backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
pinecone-client==2.2.4
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2